Add files via upload
A portion of code from my senior project, in which I developed a program alongside my team which took in multiple files of client data, generated additional data for values such as appointment locations, stored the data in a database and combined this data together to create a master list per the client's request. The original version of this file handled the file importing process and additional data generation, but as all references to the client company's name have been replaced with {name} to protect proprietary information this version of the file would not properly run even if the other files of the software were present as well and is purely for demonstrative purposes.